backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-groq
mcp
pydantic
python-dotenv
httpx
